MOB-4487 Add loyalty disabled unit test + PR comments
