+++ uisdk/src/androidTest/java/com/karhoo/uisdk/booking/checkout/bookingconfirmation/BookingConfirmationScreenshotTests.kt
@@ -1,6 +1,8 @@
 package com.karhoo.uisdk.booking.checkout.bookingconfirmation
 
+import android.app.Activity
 import android.content.Intent
+import android.os.Bundle
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
 import androidx.test.platform.app.InstrumentationRegistry
@@ -8,21 +10,26 @@ import com.github.tomakehurst.wiremock.common.ConsoleNotifier
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration
 import com.github.tomakehurst.wiremock.extension.responsetemplating.ResponseTemplateTransformer
 import com.github.tomakehurst.wiremock.junit.WireMockRule
-import com.karhoo.sdk.api.model.AuthenticationMethod
+import com.karhoo.sdk.api.model.*
+import com.karhoo.sdk.api.network.request.QuoteQTA
 import com.karhoo.uisdk.KarhooUISDK
 import com.karhoo.uisdk.R
 import com.karhoo.uisdk.common.ScreenshotTest
 import com.karhoo.uisdk.common.preferences
 import com.karhoo.uisdk.common.serverRobot
 import com.karhoo.uisdk.common.testrunner.UiSDKTestConfig
+import com.karhoo.uisdk.screen.booking.domain.address.JourneyDetails
 import com.karhoo.uisdk.util.TestData
 import com.karhoo.uisdk.util.TestSDKConfig
+import org.joda.time.DateTime
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import java.net.HttpURLConnection
+import java.text.SimpleDateFormat
+import java.util.*
 
 @RunWith(AndroidJUnit4::class)
 @LargeTest
@@ -31,10 +38,10 @@ class BookingConfirmationScreenshotTests :
 
     /**
      * In order to generate locally screenshots, you need to use the -Precord argument
-     * gradlew fulldebugexecuteScreenshotTests -Precord -Pandroid.testInstrumentationRunnerArguments.class=com.karhoo.uisdk.quotes.screenshot.QuotesScreenshotTests
+     * gradlew fulldebugexecuteScreenshotTests -Precord -Pandroid.testInstrumentationRunnerArguments.class=com.karhoo.uisdk.booking.checkout.bookingconfirmation.BookingConfirmationScreenshotTests
      *
      * When you want to compare the screenshots after some UI changes, just run
-     * gradlew fulldebugexecuteScreenshotTests -Pandroid.testInstrumentationRunnerArguments.class=com.karhoo.uisdk.quotes.screenshot.QuotesScreenshotTests
+     * gradlew fulldebugexecuteScreenshotTests -Pandroid.testInstrumentationRunnerArguments.class=com.karhoouisdk.booking.checkout.bookingconfirmation.BookingConfirmationScreenshotTests
      *
      * In order to run all screenshot tests, remove the -P argument following the gradle task
      * gradlew fulldebugexecuteScreenshotTests
@@ -77,6 +84,41 @@ class BookingConfirmationScreenshotTests :
         wireMockRule.resetAll()
     }
 
+    private fun startBookingConfirmationActivityWithDefaultValues(): Activity {
+        val bundle = Bundle()
+        bundle.putParcelable(SCREENSHOT_TEST_JOURNEY_DETAILS, journeyDetails)
+        bundle.putParcelable(SCREENSHOT_TEST_QUOTE, QUOTE)
+        bundle.putBoolean(SCREENSHOT_TEST_LOYALTY_VISIBILITY, true)
+
+        val intent = Intent(
+            InstrumentationRegistry.getInstrumentation().targetContext,
+            MockBookingConfirmationBaseActivity::class.java
+        )
+
+        intent.putExtras(bundle)
+
+        return startActivity(
+            intent
+        )
+    }
+
+    private fun startBookingConfirmationActivityWithSecondaryValues(): Activity {
+        val bundle = Bundle()
+        bundle.putParcelable(SCREENSHOT_TEST_JOURNEY_DETAILS, journeyDetails)
+        bundle.putParcelable(SCREENSHOT_TEST_QUOTE, QUOTE)
+        bundle.putBoolean(SCREENSHOT_TEST_LOYALTY_VISIBILITY, false)
+
+        val intent = Intent(
+            InstrumentationRegistry.getInstrumentation().targetContext,
+            MockBookingConfirmationBaseActivity::class.java
+        )
+
+        intent.putExtras(bundle)
+
+        return startActivity(
+            intent
+        )
+    }
 
     /**
      * Given:   The booking confirmation opens up
@@ -85,12 +127,7 @@ class BookingConfirmationScreenshotTests :
      **/
     @Test
     fun bookingConfirmationTitleIsVisible() {
-        val activity = startActivity(
-            Intent(
-                InstrumentationRegistry.getInstrumentation().targetContext,
-                MockBookingConfirmationBaseActivity::class.java
-            )
-        )
+        val activity = startBookingConfirmationActivityWithDefaultValues()
 
         bookingConfirmation {
             scrollUpBookingConfirmation()
@@ -109,12 +146,7 @@ class BookingConfirmationScreenshotTests :
      **/
     @Test
     fun bookingConfirmationPickupPrimaryAddressCorrect() {
-        val activity = startActivity(
-            Intent(
-                InstrumentationRegistry.getInstrumentation().targetContext,
-                MockBookingConfirmationBaseActivity::class.java
-            )
-        )
+        val activity = startBookingConfirmationActivityWithDefaultValues()
 
         bookingConfirmation {
             scrollUpBookingConfirmation()
@@ -132,12 +164,7 @@ class BookingConfirmationScreenshotTests :
      **/
     @Test
     fun bookingConfirmationPickupSecondaryAddressCorrect() {
-        val activity = startActivity(
-            Intent(
-                InstrumentationRegistry.getInstrumentation().targetContext,
-                MockBookingConfirmationBaseActivity::class.java
-            )
-        )
+        val activity = startBookingConfirmationActivityWithDefaultValues()
 
         bookingConfirmation {
             scrollUpBookingConfirmation()
@@ -155,12 +182,7 @@ class BookingConfirmationScreenshotTests :
      **/
     @Test
     fun bookingConfirmationDestinationPrimaryAddressCorrect() {
-        val activity = startActivity(
-            Intent(
-                InstrumentationRegistry.getInstrumentation().targetContext,
-                MockBookingConfirmationBaseActivity::class.java
-            )
-        )
+        val activity = startBookingConfirmationActivityWithDefaultValues()
 
         bookingConfirmation {
             scrollUpBookingConfirmation()
@@ -178,12 +200,7 @@ class BookingConfirmationScreenshotTests :
      **/
     @Test
     fun bookingConfirmationDestinationSecondaryAddressCorrect() {
-        val activity = startActivity(
-            Intent(
-                InstrumentationRegistry.getInstrumentation().targetContext,
-                MockBookingConfirmationBaseActivity::class.java
-            )
-        )
+        val activity = startBookingConfirmationActivityWithDefaultValues()
 
         bookingConfirmation {
             scrollUpBookingConfirmation()
@@ -201,12 +218,7 @@ class BookingConfirmationScreenshotTests :
      **/
     @Test
     fun bookingConfirmationTimeCorrect() {
-        val activity = startActivity(
-            Intent(
-                InstrumentationRegistry.getInstrumentation().targetContext,
-                MockBookingConfirmationBaseActivity::class.java
-            )
-        )
+        val activity = startBookingConfirmationActivityWithDefaultValues()
 
         bookingConfirmation {
             scrollUpBookingConfirmation()
@@ -224,12 +236,7 @@ class BookingConfirmationScreenshotTests :
      **/
     @Test
     fun bookingConfirmationDateCorrect() {
-        val activity = startActivity(
-            Intent(
-                InstrumentationRegistry.getInstrumentation().targetContext,
-                MockBookingConfirmationBaseActivity::class.java
-            )
-        )
+        val activity = startBookingConfirmationActivityWithDefaultValues()
 
         bookingConfirmation {
             scrollUpBookingConfirmation()
@@ -247,12 +254,7 @@ class BookingConfirmationScreenshotTests :
      **/
     @Test
     fun bookingConfirmationPriceCorrect() {
-        val activity = startActivity(
-            Intent(
-                InstrumentationRegistry.getInstrumentation().targetContext,
-                MockBookingConfirmationBaseActivity::class.java
-            )
-        )
+        val activity = startBookingConfirmationActivityWithDefaultValues()
 
         bookingConfirmation {
             scrollUpBookingConfirmation()
@@ -270,12 +272,7 @@ class BookingConfirmationScreenshotTests :
      **/
     @Test
     fun bookingConfirmationEstimatedFareTypeCorrect() {
-        val activity = startActivity(
-            Intent(
-                InstrumentationRegistry.getInstrumentation().targetContext,
-                MockBookingConfirmationBaseActivity::class.java
-            )
-        )
+        val activity = startBookingConfirmationActivityWithDefaultValues()
 
         bookingConfirmation {
             scrollUpBookingConfirmation()
@@ -293,12 +290,7 @@ class BookingConfirmationScreenshotTests :
      **/
     @Test
     fun bookingConfirmationEstimatedLoyaltyVisible() {
-        val activity = startActivity(
-            Intent(
-                InstrumentationRegistry.getInstrumentation().targetContext,
-                MockBookingConfirmationBaseActivity::class.java
-            )
-        )
+        val activity = startBookingConfirmationActivityWithDefaultValues()
 
         bookingConfirmation {
             scrollUpBookingConfirmation()
@@ -309,5 +301,134 @@ class BookingConfirmationScreenshotTests :
         compareScreenshot(activity)
     }
 
+    /**
+     * Given:   The booking confirmation opens up
+     * When:    It has finished the transition on screen
+     * Then:    The loyalty is not visible
+     **/
+    @Test
+    fun bookingConfirmationEstimatedLoyaltyNotVisible() {
+        val activity = startBookingConfirmationActivityWithSecondaryValues()
 
+        bookingConfirmation {
+            scrollUpBookingConfirmation()
+            checkLoyaltyNotVisible()
+            shortSleep()
+        }
+
+        compareScreenshot(activity)
+    }
+
+    /**
+     * Given:   The booking confirmation opens up
+     * When:    It has finished the transition on screen
+     * Then:    The add to calendar should be visible
+     **/
+    @Test
+    fun bookingConfirmationAddToCalendarVisible() {
+        val activity = startBookingConfirmationActivityWithDefaultValues()
+
+        bookingConfirmation {
+            scrollUpBookingConfirmation()
+            addToCalendarVisibility()
+            shortSleep()
+        }
+
+        compareScreenshot(activity)
+    }
+
+    companion object {
+        val TRIP_POSITION_PICKUP = Position(
+            latitude = 51.523766,
+            longitude = -0.1375291
+        )
+
+        val TRIP_POSITION_DROPOFF = Position(
+            latitude = 51.514432,
+            longitude = -0.1585557
+        )
+
+        val TRIP_LOCATION_INFO_PICKUP = LocationInfo(
+            address = Address(
+                displayAddress = "221B Baker St, Marylebone, London NW1 6XE, UK",
+                buildingNumber = "221B",
+                streetName = "Baker St",
+                city = "Marleybone",
+                postalCode = "NW1 6XE",
+                region = "London",
+                countryCode = "UK"
+            ),
+            position = TRIP_POSITION_PICKUP,
+            placeId = "ChIJEYJiM88adkgR4SKDqHd2XUQ",
+            timezone = "Europe/London",
+            poiType = Poi.NOT_SET
+        )
+
+        val TRIP_LOCATION_INFO_DROPOFF = LocationInfo(
+            address =  Address(
+                displayAddress = "368 Oxford St, London W1D 1LU, UK",
+                buildingNumber = "368",
+                streetName = "Oxford St",
+                city = "London",
+                postalCode = "W1D 1LU",
+                countryCode = "UK"
+            ),
+            position = TRIP_POSITION_DROPOFF,
+            placeId = "ChIJyWu2IisbdkgRHIRWuD0ANfM",
+            timezone = "Europe/London",
+            poiType = Poi.NOT_SET
+        )
+        val QUOTE_PRICE = QuotePrice(
+            currencyCode = "GBP",
+            highPrice = 577,
+            lowPrice = 577
+        )
+
+        val QUOTE_FLEET = Fleet(
+            id = "52123bd9-cc98-4b8d-a98a-122446d69e79",
+            name = "iCabbi [Sandbox]",
+            logoUrl = "https://cdn.karhoo.com/d/images/logos/52123bd9-cc98-4b8d-a98a-122446d69e79.png",
+            description = "Some fleet description",
+            phoneNumber = "+447904839920",
+            termsConditionsUrl = "http://www.google.com"
+        )
+
+        val QUOTE_VEHICLE = QuoteVehicle(
+            vehicleClass = "Electric",
+            vehicleQta = QuoteQTA(highMinutes = 30, lowMinutes = 1),
+            luggageCapacity = 2,
+            vehicleType = "standard",
+            vehicleTags = arrayListOf("taxi, hybrid"),
+            passengerCapacity = 2
+        )
+
+        val QUOTE = Quote(
+            id = "NTIxMjNiZDktY2M5OC00YjhkLWE5OGEtMTIyNDQ2ZDY5ZTc5O3NhbG9vbg==",
+            quoteType = QuoteType.ESTIMATED,
+            quoteSource = QuoteSource.FLEET,
+            price = QUOTE_PRICE,
+            fleet = QUOTE_FLEET,
+            pickupType = PickupType.CURBSIDE,
+            vehicle = QUOTE_VEHICLE
+        )
+
+        fun getDate(dateScheduled: String): Date {
+            val formatter = SimpleDateFormat("yyyy-MM-dd'T'HH:mm").apply {
+                timeZone = TimeZone.getTimeZone("UTC")
+            }
+            return formatter.parse(dateScheduled)
+        }
+
+        val SCHEDULED_DATE = getDate("2019-07-31T12:35:00Z")
+
+        val journeyDetails = JourneyDetails(
+            TRIP_LOCATION_INFO_PICKUP,
+            TRIP_LOCATION_INFO_DROPOFF,
+            DateTime(SCHEDULED_DATE)
+        )
+
+        const val SCREENSHOT_TEST_JOURNEY_DETAILS = "SCREENSHOT_TEST_JOURNEY_DETAILS"
+        const val SCREENSHOT_TEST_QUOTE = "SCREENSHOT_TEST_QUOTE"
+        const val SCREENSHOT_TEST_LOYALTY_VISIBILITY = "SCREENSHOT_TEST_LOYALTY_VISIBILITY"
+    }
 }
